#182 Implement IPU based population synthesis
Fix overflow due to cast from int to short while extending the maximum
duration to the one in PatternActivity.

+++ mobitopp/src/main/java/edu/kit/ifv/mobitopp/simulation/activityschedule/linkedlist/ActivityAsLinkedListElement.java
@@ -55,7 +55,7 @@ public class ActivityAsLinkedListElement
 		float durationFlexibiliy
 	)
   { 
-		super(oid, (byte) activityNrOfWeek, activityType, startDate, (short) duration, (short) observedTripDuration,
+		super(oid, (byte) activityNrOfWeek, activityType, startDate, duration, observedTripDuration,
 				startFlexibility, endFlexibility, durationFlexibiliy);
   }
 }
